requirements.txt
python-dotenv
geopy
requests
pytz
newsapi-python